+++ R/ReleaseQuestions.R
@@ -1,5 +1,6 @@
 release_questions <- function() {
   c(
-    "Are vignettes up to date?"
+    "Are vignettes up to date?",
+    "Have you updated .zenodo.json?"
   )
 }