+++ NEWS.md
@@ -1,5 +1,6 @@
 # Ternary v0.1.1
  - Added colour-blind compatible colour palettes `cbPalette8` and `cbPalette15`.
+ - Support lab.font in TernaryPlot
 
 # Ternary v0.1.0
  - Generates ternary plots with vertical left axis.