+++ NEWS.md
@@ -1,5 +1,6 @@
 # Ternary vNEXT
  - Add function JoinTheDots to plot lines and points simultaneously.
+ - Support xlim & ylim for TernaryPlot
 
 # Ternary v0.1.1
  - Added colour-blind compatible colour palettes `cbPalette8` and `cbPalette15`.