+++ NEWS.md
@@ -1,6 +1,7 @@
 # Ternary vNEXT
  - Add function JoinTheDots to plot lines and points simultaneously.
  - Support xlim & ylim for TernaryPlot
+ - Improveme documentation
 
 # Ternary v0.1.1
  - Added colour-blind compatible colour palettes `cbPalette8` and `cbPalette15`.