+++ NEWS.md
@@ -1,6 +1,7 @@
 # Ternary vNEXT
  - Add function JoinTheDots to plot lines and points simultaneously.
  - Support xlim & ylim for TernaryPlot
+ - Improved control over labelling of plot (clab.rotate parameter)
  - Improveme documentation
 
 # Ternary v0.1.1